requirements.txt
# Core dependencies for Founder Dataset Generation Pipeline
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# LLM Provider
google-generativeai>=0.3.0
